Refactor: file formats
