+++ src/fileformats/file_format_registry.cpp
