+++ src/fileformats/ocd_file_export.h
@@ -23,7 +23,7 @@
 #ifndef OPENORIENTEERING_OCD_FILE_EXPORT_H
 #define OPENORIENTEERING_OCD_FILE_EXPORT_H
 
-#include "../file_import_export.h"
+#include "file_import_export.h"
 
 
 /**
