+++ src/fileformats/ocd_file_import.cpp
@@ -34,25 +34,25 @@
 #include "ocd_types_v10.h"
 #include "ocd_types_v11.h"
 #include "ocd_types_v12.h"
-#include "../core/crs_template.h"
-#include "../core/map_color.h"
-#include "../core/map_view.h"
-#include "../core/georeferencing.h"
-#include "../file_format_ocad8.h"
-#include "../file_format_ocad8_p.h"
-#include "../map.h"
-#include "../object_text.h"
-#include "../settings.h"
-#include "../symbol_area.h"
-#include "../symbol_combined.h"
-#include "../symbol_line.h"
-#include "../symbol_point.h"
-#include "../symbol_text.h"
-#include "../template.h"
-#include "../template_image.h"
-#include "../template_map.h"
-#include "../util.h"
-#include "../util/encoding.h"
+#include "core/crs_template.h"
+#include "core/map_color.h"
+#include "core/map_view.h"
+#include "core/georeferencing.h"
+#include "ocad8_file_format.h"
+#include "ocad8_file_format_p.h"
+#include "map.h"
+#include "object_text.h"
+#include "settings.h"
+#include "symbol_area.h"
+#include "symbol_combined.h"
+#include "symbol_line.h"
+#include "symbol_point.h"
+#include "symbol_text.h"
+#include "template.h"
+#include "template_image.h"
+#include "template_map.h"
+#include "util.h"
+#include "util/encoding.h"
 
 
 namespace {
