+++ src/gdal/gdal_settings_page.cpp
@@ -29,7 +29,7 @@
 #include "gdal_manager.h"
 #include "ogr_file_format.h"
 #include "../util_gui.h"
-#include "../file_format_registry.h"
+#include "../fileformats/file_format_registry.h"
 #include "../util/scoped_signals_blocker.h"
