+++ src/gdal/ogr_file_format.h
@@ -20,7 +20,7 @@
 #ifndef OPENORIENTEERING_OGR_FILE_FORMAT_H
 #define OPENORIENTEERING_OGR_FILE_FORMAT_H
 
-#include "../file_format.h"
+#include "../fileformats/file_format.h"
 
 /**
  * A FileFormat for geospatial vector data supported by OGR.
