+++ src/gdal/ogr_file_format_p.h
@@ -30,7 +30,7 @@
 #include <ogr_srs_api.h>
 
 #include "../core/map_coord.h"
-#include "../file_import_export.h"
+#include "../fileformats/file_import_export.h"
 #include "../symbol.h"
 
 class AreaSymbol;
