+++ src/gui/main_window.cpp
@@ -49,13 +49,13 @@
 #include "home_screen_controller.h"
 #include "settings_dialog.h"
 #include "../core/map_view.h"
-#include "../file_format_registry.h"
-#include "../file_import_export.h"
+#include "../fileformats/file_format_registry.h"
+#include "../fileformats/file_import_export.h"
 #include "../map.h"
 #include "../map_dialog_new.h"
 #include "../map_editor.h"
 #include "../mapper_resource.h"
-#include "../file_format.h"
+#include "../fileformats/file_format.h"
 #include "../settings.h"
 #include "../symbol.h"
 #include "../undo_manager.h"
