+++ src/map.cpp
@@ -37,9 +37,9 @@
 #include "core/map_color.h"
 #include "core/map_printer.h"
 #include "core/map_view.h"
-#include "file_format_ocad8.h"
-#include "file_format_registry.h"
-#include "file_import_export.h"
+#include "fileformats/ocad8_file_format.h"
+#include "fileformats/file_format_registry.h"
+#include "fileformats/file_import_export.h"
 #include "map_editor.h"
 #include "map_part.h"
 #include "object_undo.h"
