+++ src/symbol.h
@@ -27,7 +27,7 @@
 #include <QRgb>
 
 #include "core/map_coord.h"
-#include "file_format.h"
+#include "fileformats/file_format.h"
 
 QT_BEGIN_NAMESPACE
 class QIODevice;
