+++ src/util/xml_stream_util.cpp
@@ -25,8 +25,8 @@
 #include <QTextStream>
 
 #include "../core/map_coord.h"
-#include "../file_format_xml.h"
-#include "../file_import_export.h"
+#include "../fileformats/xml_file_format.h"
+#include "../fileformats/file_import_export.h"
 
 
 //### XmlRecoveryHelper ###
