+++ test/file_format_t.cpp
@@ -24,9 +24,9 @@
 #include "../src/core/map_color.h"
 #include "../src/core/map_grid.h"
 #include "../src/core/map_printer.h"
-#include "../src/file_import_export.h"
-#include "../src/file_format_ocad8.h"
-#include "../src/file_format_registry.h"
+#include "../src/fileformats/file_import_export.h"
+#include "../src/fileformats/ocad8_file_format.h"
+#include "../src/fileformats/file_format_registry.h"
 #include "../src/global.h"
 #include "../src/mapper_resource.h"
 #include "../src/object.h"
